+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -104,7 +104,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kh5206fcyyeek0x4v41bna6e"
+        "conv_01kh5hhzhcs2depjr6hxpvpmj8"
     )
     r"""Unique conversation identifier with `conv_` prefix."""