+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -199,7 +199,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kh5206few1spvqzbfx9hay2c"
+        "conv_01kh5hhzhet9wyb9srr3k3nt1d"
     )
     r"""Unique conversation identifier with `conv_` prefix."""