+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH52070WG3C2CTN75FVVN6DH"
+        "01KH5HJ01AM0ZEX28XSEP736NX"
     )
     r"""The unique identifier of the data source"""